refactor: moved widgets to textfields_demo.dart
